% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Forest_seg.R
\name{analyze_forest_canopy}
\alias{analyze_forest_canopy}
\title{Analyze forest canopy (internal)}
\usage{
analyze_forest_canopy(
  AGB_def,
  Forest_floor,
  voxel_size = 0.1,
  min_height = 1.5,
  coverage_method = "mean_normalized",
  area_of_interest = NULL,
  output_path = tempdir()
)
}
\description{
Analyze forest canopy (internal)
}
\keyword{internal}
