% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Forest_seg.R
\name{calculate_coverage_degree}
\alias{calculate_coverage_degree}
\title{Calculate coverage degree (internal)}
\usage{
calculate_coverage_degree(voxel_data, method = "mean_normalized")
}
\description{
Calculate coverage degree (internal)
}
\keyword{internal}
