% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Forest_seg.R
\name{calculate_crown_base_heights}
\alias{calculate_crown_base_heights}
\title{Calculate crown base heights (internal)}
\usage{
calculate_crown_base_heights(tree_ids, plot_metrics, AGB_def)
}
\description{
Calculate crown base heights (internal)
}
\keyword{internal}
