% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Forest_seg.R
\name{calculate_vertical_profile}
\alias{calculate_vertical_profile}
\title{Calculate vertical profile (internal)}
\usage{
calculate_vertical_profile(
  voxel_data,
  w_column,
  voxel_size,
  height_interval = 1
)
}
\description{
Calculate vertical profile (internal)
}
\keyword{internal}
