% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Forest_seg.R
\name{extract_forest_floor}
\alias{extract_forest_floor}
\title{Extract forest floor (internal)}
\usage{
extract_forest_floor(a, soil_dim, th = 20, N = 500, output_path, plot)
}
\description{
Extract forest floor (internal)
}
\keyword{internal}
