% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Forest_seg.R
\name{filter_trees_by_point_count}
\alias{filter_trees_by_point_count}
\title{Filter trees by point count (internal)}
\usage{
filter_trees_by_point_count(plot_metrics, woodpoint)
}
\description{
Filter trees by point count (internal)
}
\keyword{internal}
