% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PiC-package.R
\docType{package}
\name{PiC-package}
\alias{PiC}
\alias{PiC-package}
\title{PiC: Interactive Processing and Segmentation of Forest TLS Point-Cloud Data}
\description{
Tools for the processing, segmentation, and analysis of terrestrial laser scanning (TLS and MLS) forest point-cloud data. The package provides fast voxel-based processing, classification of point clouds into forest floor, understory, canopy, and woody components, and algorithms for single-tree analysis and structural characterization. Methods are designed to handle large and dense point-cloud datasets efficiently, supporting applications in forest structure assessment, connectivity analysis, and fire-risk evaluation. Input data are provided as '.xyz', '.txt', '.las', or '.laz' point-cloud files. For methodological details, see Ferrara and Arrizza (2025) \url{https://hdl.handle.net/20.500.14243/533471} and Ferrara et al. (2018) \doi{10.1016/j.agrformet.2018.04.008}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/rupppy/PiC}
  \item Report bugs at \url{https://github.com/rupppy/PiC/issues}
}

}
\author{
\strong{Maintainer}: Roberto Ferrara \email{roberto.ferrara@cnr.it} (\href{https://orcid.org/0009-0000-3627-6867}{ORCID})

Other contributors:
\itemize{
  \item Stefano Arrizza \email{stefano.arrizza@cnr.it} (\href{https://orcid.org/0009-0009-2290-3650}{ORCID}) [contributor]
}

}
\keyword{internal}
