% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pic.R
\name{pic}
\alias{pic}
\title{Take a pic of your package}
\usage{
pic(pkg, type = c("selfie", "friends", "followers"))
}
\arguments{
\item{pkg}{target package (char/scalar)}
}
\description{
Take a pic of your package
}
