% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_deps_edges.R
\name{pkg_deps_edges}
\alias{pkg_deps_edges}
\title{Get all package dependencies/dependents}
\usage{
pkg_deps_edges(pkg, direction = c("in", "out"))
}
\arguments{
\item{pkg}{package name}

\item{direction}{dependency direction (in/out)}
}
\description{
Get all package dependencies/dependents
}
