% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pic.R
\name{pkg_deps_pic}
\alias{pkg_deps_pic}
\title{Take a package dependency (i.e., friend/follower) pic}
\usage{
pkg_deps_pic(pkg, direction)
}
\arguments{
\item{pkg}{package name}

\item{direction}{dependency direction (in/out)}
}
\description{
Take a package dependency (i.e., friend/follower) pic
}
