% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_fun_edges.R
\name{pkg_fun_edges}
\alias{pkg_fun_edges}
\title{Get dependent functions within the package}
\usage{
pkg_fun_edges(fun, funs)
}
\arguments{
\item{fun}{function to find dependents for}

\item{funs}{list of possible dependents}
}
\description{
Get dependent functions within the package
}
