% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_funs.R
\name{pkg_funs}
\alias{pkg_funs}
\title{Get a list of all functions within the package}
\usage{
pkg_funs(pkg)
}
\arguments{
\item{pkg}{target package (char/scalar)}
}
\description{
Get a list of all functions within the package
}
