% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_vars.R
\name{pkg_vars}
\alias{pkg_vars}
\title{Get a list of all defined variables within the package}
\usage{
pkg_vars(pkg)
}
\arguments{
\item{pkg}{target package (char/scalar)}
}
\description{
Get a list of all defined variables within the package
}
