% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pic.R
\name{plot.pic}
\alias{plot.pic}
\title{Draw a package pic}
\usage{
\method{plot}{pic}(x, type = c("selfie", "friends", "followers"),
  show_names = TRUE, show_edges = TRUE)
}
\arguments{
\item{x}{pic object}

\item{type}{pic type (selfie/friends/followers)}

\item{show_names}{show names (logical/scalar)}

\item{show_edges}{show dependencies (logical/scalar)}
}
\description{
Draw a package pic
}
