% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Forest_seg.R
\name{segment_wood}
\alias{segment_wood}
\title{Segment wood (internal)}
\usage{
segment_wood(
  AGB,
  dim,
  th,
  eps,
  mpts,
  h_tree,
  N,
  R,
  Vox_print,
  WoodVox_print,
  plot,
  output_path
)
}
\description{
Segment wood (internal)
}
\keyword{internal}
