% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Forest_seg.R
\name{separate_foliage}
\alias{separate_foliage}
\title{Separate foliage (internal)}
\usage{
separate_foliage(AGB, woodpoint, plot, eps, mpts, output_path)
}
\description{
Separate foliage (internal)
}
\keyword{internal}
