% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wood_seg.R
\name{Wood_seg}
\alias{Wood_seg}
\title{Wood voxels segmentation}
\arguments{
\item{filename}{- Output file prefix}

\item{a}{- AGB voxelized input file}

\item{eps}{- size (radius) of the epsilon neighborhood - Default = 1}

\item{mpts}{- number of minimum points required in the eps neighborhood for core points (including the point itself) - Default = 4}

\item{N}{- Minimum number of voxel in a wood cluster - Default = 1000}

\item{R}{- R = Standard deviation * Proportion of Variance - Default = 30}
}
\value{
One file (.txt) output - Wood voxels (vox)
}
\description{
Point cloud segmentation to identify wood voxels
}
