% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Forest_floor3.R
\name{Floseg}
\alias{Floseg}
\title{Forest floor segmentation}
\usage{
Floseg(a, filename="XXX", Soil_dim = 30, th = 20, N=500)
}
\arguments{
\item{a}{- Input file (.xyz)}

\item{filename}{- Output file prefix}

\item{Soil_dim}{- Voxel dimension (cm) for forest floor segmentation - Default = 30}

\item{th}{- Minimum number of point to generate a voxel. Default = 20}

\item{N}{- Minimum number of voxel to generate a cluster. Default = 500}
}
\value{
2 files (.txt) output. 1. Forest floor pointcolud; 2. AGB pointcloud
}
\description{
Segments the input .xyz pointcloud file into different forestry layers: forest floor and above ground biomass.
}
