% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Floseg.R
\name{Floseg}
\alias{Floseg}
\title{Forest floor segmentation}
\usage{
Floseg(a, filename="XXX", soil_dim = 0.3, th = 20, N=500, output_path = tempdir())
}
\arguments{
\item{a}{- Input file (.xyz)}

\item{filename}{- Output file prefix}

\item{soil_dim}{- Voxel dimension (m) for forest floor segmentation - Default = 0.30}

\item{th}{- Minimum number of point to generate a voxel. Default = 20}

\item{N}{- Minimum number of voxel to generate a cluster. Default = 500}

\item{output_path}{Directory in cui scrivere i file di output. Default = tempdir()}
}
\value{
2 files (.txt) output. 1. Forest floor pointcolud; 2. AGB pointcloud
}
\description{
Segments the input .xyz pointcloud file into different forestry layers: forest floor and above ground biomass.
}
