\name{PIC-package}
\alias{PIC-package}
\alias{PIC}
\docType{package}
\title{
\packageTitle{PIC}
}
\description{
\packageDescription{PIC}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{PIC}
\packageIndices{PIC}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{PIC}

Maintainer: \packageMaintainer{PIC}
}
\references{
~~ Literature or other references for background information ~~
}

\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
