% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PiC-package.R
\docType{package}
\name{PiC-package}
\alias{PiC}
\alias{PiC-package}
\title{PiC: Pointcloud Interactive Computation}
\description{
Provides advanced algorithms for analyzing pointcloud data from terrestrial laser scanner in forestry applications. Key features include fast voxelization of large datasets; segmentation of point clouds into forest floor, understorey, canopy, and wood components. The package enables efficient processing of large-scale forest pointcloud data, offering insights into forest structure, connectivity, and fire risk assessment. Algorithms to analyze pointcloud data (.xyz input file). For more details, see Ferrara & Arrizza (2025) \url{https://hdl.handle.net/20.500.14243/533471}. For single tree segmentation details, see Ferrara et al. (2018) \doi{10.1016/j.agrformet.2018.04.008}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/rupppy/PiC}
  \item Report bugs at \url{https://github.com/rupppy/PiC/issues}
}

}
\author{
\strong{Maintainer}: Roberto Ferrara \email{roberto.ferrara@cnr.it} (\href{https://orcid.org/0009-0000-3627-6867}{ORCID})

Other contributors:
\itemize{
  \item Stefano Arrizza (\href{https://orcid.org/0009-0009-2290-3650}{ORCID}) [contributor]
}

}
\keyword{internal}
