% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Run_app.R
\name{run_PiC}
\alias{run_PiC}
\title{Launch PiC Shiny App}
\usage{
run_PiC()
}
\value{
No return value, called for side effects (launches Shiny app)
}
\description{
Launch the Shiny app for interactive 3D point cloud processing.
}
\details{
This function launches an interactive web application for analyzing
  forest point cloud data. The app requires additional packages that are not
  installed by default. If these packages are missing, you will be prompted
  to install them.
}
\examples{
\dontrun{
# Launch the interactive app
run_PiC()
}
}
