% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Single_tree_seg.R
\name{SegOne}
\alias{SegOne}
\title{Single Tree wood leaf segmentation}
\arguments{
\item{dimVox}{- voxel dimension in cm - Default = 2}

\item{th}{- Minimum number of points to generate a voxel - Default = 2}

\item{filename}{- Output file prefix}

\item{a}{- AGB voxelized input file}

\item{eps}{- size (radius) of the epsilon neighborhood - Default = 1}

\item{mpts}{- number of minimum points required in the eps neighborhood for core points (including the point itself) - Default = 4}

\item{N}{- Minimum number of voxel in a wood cluster - Default = 1000}

\item{R}{- R = Standard deviation * Proportion of Variance - Default = 30}
}
\value{
Two file (.txt) in output - Wood points and non wood points
}
\description{
Wood - leaf segmentation of single tree
}
